Change the query to accpect host define by user
I change the Query class: the user of this API can define the ip where he want to apply the query.
For me this is importante because, if the pc was more than one network interface the query using the InetAddress.getLocalHost() can access to the network that doens't comunication with other services. So this away the user can define the ip of network interface that want